+++ repos/re-theme/src/theme/index.js
@@ -1,4 +1,4 @@
-export * from './buildTheme'
+export * from './getDefaultPlatforms'
 export * from './default'
 export * from './themeEvent'
 export * from './withTheme'
